fix(fft): int64 sparse delta indices, bounds-check patches, exit on a dead engine
An 80-step Gemma-4-e2b full fine-tune died at step 21: the sampler's
EngineCore hit "index_copy_(): index out of bounds" while applying the
delta in place. The trainer wrote delta indices as int32, and Gemma 4's
per-layer embedding table has 2,348,810,240 elements, 201M past the int32
limit. The first step that touched a row for a token id above ~239,674
produced 115 negative indices, which the CUDA kernel asserted on. Deltas
1-20 had happened not to reach those rows. Qwen tensors never exceed 2^31
elements, so only Gemma (and other large-vocab, large-hidden models) hit it.

- Trainer writes indices as int64 (its own parameter shadow update used the
  same indices; a negative one there silently wrote to the tensor's tail).
- The sampler bounds-checks every patch on the host before touching the
  device, on both the in-place and the CPU-snapshot path, and names the
  parameter in the error instead of surfacing an opaque device assert.
- The in-place loop labels patches by their HF name; str(gpu_param) was
  formatting entire tensors on the device on every layer.
- A dead EngineCore no longer leaves the worker Running with hung clients:
  popped-but-unserved requests are failed explicitly and the process exits
  non-zero so the pod restarts. The run above hung for the client's full
  7200s timeout before anyone noticed.

+++ src/training/fft_trainer_worker.py
@@ -245,12 +245,15 @@ class FFTTrainingWorker(BaseTrainerWorker):
       indices_list = []
       values_list = []
 
+    # int64 indices: a flat index into a tensor with more than 2**31 elements
+    # (Gemma 4's per-layer embedding table is 2.35e9) does not fit an int32, and
+    # a wrapped negative index made the sampler's index_copy_ assert mid-run.
     if indices_list:
-      indices_flat = torch.cat(indices_list).to(torch.int32).contiguous()
+      indices_flat = torch.cat(indices_list).to(torch.int64).contiguous()
       values_flat = torch.cat(values_list).contiguous()
     else:
       fallback_dtype = next(self.model.parameters()).dtype if self.model else torch.float32
-      indices_flat = torch.empty(0, dtype=torch.int32, device="cpu")
+      indices_flat = torch.empty(0, dtype=torch.int64, device="cpu")
       values_flat = torch.empty(0, dtype=fallback_dtype, device="cpu")
 
     layer_lengths_tensor = torch.tensor(layer_lengths_list, dtype=torch.int64, device="cpu")
@@ -469,7 +472,7 @@ class FFTTrainingWorker(BaseTrainerWorker):
         diff_mask = param.data.view(-1).ne(prev_gpu.view(-1))
         indices = diff_mask.nonzero(as_tuple=True)[0]
         if indices.numel() > 0:
-          idx_cpu = indices.to(torch.int32).contiguous().cpu()
+          idx_cpu = indices.to(torch.int64).contiguous().cpu()
           val_cpu = param.data.view(-1)[diff_mask].contiguous().cpu()
           layer_names_list.append(name)
           indices_list.append(idx_cpu)
